Move envelope editor functions and types to separate source file
This purely moves functions and types from the source files `editor.cpp` and `popups.cpp` to the new source file `envelope_editor.cpp`.
