+++ CMakeLists.txt
@@ -2648,6 +2648,7 @@ if(CLIENT)
     editor_ui.cpp
     editor_ui.h
     enums.h
+    envelope_editor.cpp
     explanations.cpp
     explanations.h
     file_browser.cpp
