+++ src/game/editor/popups.cpp
@@ -1397,289 +1397,6 @@ CUi::EPopupMenuFunctionResult CEditor::PopupPoint(void *pContext, CUIRect View, 
 	return CUi::POPUP_KEEP_OPEN;
 }
 
-CUi::EPopupMenuFunctionResult CEditor::PopupEnvPoint(void *pContext, CUIRect View, bool Active)
-{
-	CEditor *pEditor = static_cast<CEditor *>(pContext);
-	if(pEditor->Map()->m_SelectedEnvelope < 0 || pEditor->Map()->m_SelectedEnvelope >= (int)pEditor->Map()->m_vpEnvelopes.size())
-		return CUi::POPUP_CLOSE_CURRENT;
-
-	const float RowHeight = 12.0f;
-	CUIRect Row, Label, EditBox;
-
-	pEditor->m_ActiveEnvelopePreview = EEnvelopePreview::SELECTED;
-
-	std::shared_ptr<CEnvelope> pEnvelope = pEditor->Map()->m_vpEnvelopes[pEditor->Map()->m_SelectedEnvelope];
-
-	if(pEnvelope->GetChannels() == 4 && !pEditor->Map()->IsTangentSelected())
-	{
-		View.HSplitTop(RowHeight, &Row, &View);
-		View.HSplitTop(4.0f, nullptr, &View);
-		Row.VSplitLeft(60.0f, &Label, &Row);
-		Row.VSplitLeft(10.0f, nullptr, &EditBox);
-		pEditor->Ui()->DoLabel(&Label, "Color:", RowHeight - 2.0f, TEXTALIGN_ML);
-
-		const auto SelectedPoint = pEditor->Map()->m_vSelectedEnvelopePoints.front();
-		const int SelectedIndex = SelectedPoint.first;
-		auto *pValues = pEnvelope->m_vPoints[SelectedIndex].m_aValues;
-		const ColorRGBA Color = pEnvelope->m_vPoints[SelectedIndex].ColorValue();
-		const auto &&SetColor = [&](ColorRGBA NewColor) {
-			if(Color == NewColor && pEditor->m_ColorPickerPopupContext.m_State == EEditState::EDITING)
-				return;
-
-			static int s_Values[4];
-
-			if(pEditor->m_ColorPickerPopupContext.m_State == EEditState::START || pEditor->m_ColorPickerPopupContext.m_State == EEditState::ONE_GO)
-			{
-				for(int Channel = 0; Channel < 4; ++Channel)
-					s_Values[Channel] = pValues[Channel];
-			}
-
-			pEnvelope->m_vPoints[SelectedIndex].SetColorValue(NewColor);
-
-			if(pEditor->m_ColorPickerPopupContext.m_State == EEditState::END || pEditor->m_ColorPickerPopupContext.m_State == EEditState::ONE_GO)
-			{
-				std::vector<std::shared_ptr<IEditorAction>> vpActions(4);
-
-				for(int Channel = 0; Channel < 4; ++Channel)
-				{
-					vpActions[Channel] = std::make_shared<CEditorActionEnvelopeEditPoint>(pEditor->Map(), pEditor->Map()->m_SelectedEnvelope, SelectedIndex, Channel, CEditorActionEnvelopeEditPoint::EEditType::VALUE, s_Values[Channel], f2fx(NewColor[Channel]));
-				}
-
-				char aDisplay[256];
-				str_format(aDisplay, sizeof(aDisplay), "Edit color of point %d of envelope %d", SelectedIndex, pEditor->Map()->m_SelectedEnvelope);
-				pEditor->Map()->m_EnvelopeEditorHistory.RecordAction(std::make_shared<CEditorActionBulk>(pEditor->Map(), vpActions, aDisplay));
-			}
-
-			pEditor->Map()->m_UpdateEnvPointInfo = true;
-			pEditor->Map()->OnModify();
-		};
-		static char s_ColorPickerButton;
-		pEditor->DoColorPickerButton(&s_ColorPickerButton, &EditBox, Color, SetColor);
-	}
-
-	static CLineInputNumber s_CurValueInput;
-	static CLineInputNumber s_CurTimeInput;
-
-	static float s_CurrentTime = 0;
-	static float s_CurrentValue = 0;
-
-	if(pEditor->Map()->m_UpdateEnvPointInfo)
-	{
-		pEditor->Map()->m_UpdateEnvPointInfo = false;
-
-		const auto &[CurrentTime, CurrentValue] = pEditor->Map()->SelectedEnvelopeTimeAndValue();
-
-		// update displayed text
-		s_CurValueInput.SetFloat(fx2f(CurrentValue));
-		s_CurTimeInput.SetFloat(CurrentTime.AsSeconds());
-
-		s_CurrentTime = s_CurTimeInput.GetFloat();
-		s_CurrentValue = s_CurValueInput.GetFloat();
-	}
-
-	View.HSplitTop(RowHeight, &Row, &View);
-	Row.VSplitLeft(60.0f, &Label, &Row);
-	Row.VSplitLeft(10.0f, nullptr, &EditBox);
-	pEditor->Ui()->DoLabel(&Label, "Value:", RowHeight - 2.0f, TEXTALIGN_ML);
-	pEditor->DoEditBox(&s_CurValueInput, &EditBox, RowHeight - 2.0f, IGraphics::CORNER_ALL, "The value of the selected envelope point.");
-
-	View.HSplitTop(4.0f, nullptr, &View);
-	View.HSplitTop(RowHeight, &Row, &View);
-	Row.VSplitLeft(60.0f, &Label, &Row);
-	Row.VSplitLeft(10.0f, nullptr, &EditBox);
-	pEditor->Ui()->DoLabel(&Label, "Time (in s):", RowHeight - 2.0f, TEXTALIGN_ML);
-	pEditor->DoEditBox(&s_CurTimeInput, &EditBox, RowHeight - 2.0f, IGraphics::CORNER_ALL, "The time of the selected envelope point.");
-
-	if(pEditor->Input()->KeyIsPressed(KEY_RETURN) || pEditor->Input()->KeyIsPressed(KEY_KP_ENTER))
-	{
-		float CurrentTime = s_CurTimeInput.GetFloat();
-		float CurrentValue = s_CurValueInput.GetFloat();
-		if(!(absolute(CurrentTime - s_CurrentTime) < 0.0001f && absolute(CurrentValue - s_CurrentValue) < 0.0001f))
-		{
-			const auto &[OldTime, OldValue] = pEditor->Map()->SelectedEnvelopeTimeAndValue();
-
-			if(pEditor->Map()->IsTangentInSelected())
-			{
-				auto [SelectedIndex, SelectedChannel] = pEditor->Map()->m_SelectedTangentInPoint;
-
-				pEditor->Map()->m_EnvelopeEditorHistory.Execute(std::make_shared<CEditorActionEditEnvelopePointValue>(pEditor->Map(), pEditor->Map()->m_SelectedEnvelope, SelectedIndex, SelectedChannel, CEditorActionEditEnvelopePointValue::EType::TANGENT_IN, OldTime, OldValue, CFixedTime::FromSeconds(CurrentTime), f2fx(CurrentValue)));
-				CurrentTime = (pEnvelope->m_vPoints[SelectedIndex].m_Time + pEnvelope->m_vPoints[SelectedIndex].m_Bezier.m_aInTangentDeltaX[SelectedChannel]).AsSeconds();
-			}
-			else if(pEditor->Map()->IsTangentOutSelected())
-			{
-				auto [SelectedIndex, SelectedChannel] = pEditor->Map()->m_SelectedTangentOutPoint;
-
-				pEditor->Map()->m_EnvelopeEditorHistory.Execute(std::make_shared<CEditorActionEditEnvelopePointValue>(pEditor->Map(), pEditor->Map()->m_SelectedEnvelope, SelectedIndex, SelectedChannel, CEditorActionEditEnvelopePointValue::EType::TANGENT_OUT, OldTime, OldValue, CFixedTime::FromSeconds(CurrentTime), f2fx(CurrentValue)));
-				CurrentTime = (pEnvelope->m_vPoints[SelectedIndex].m_Time + pEnvelope->m_vPoints[SelectedIndex].m_Bezier.m_aOutTangentDeltaX[SelectedChannel]).AsSeconds();
-			}
-			else
-			{
-				auto [SelectedIndex, SelectedChannel] = pEditor->Map()->m_vSelectedEnvelopePoints.front();
-				pEditor->Map()->m_EnvelopeEditorHistory.Execute(std::make_shared<CEditorActionEditEnvelopePointValue>(pEditor->Map(), pEditor->Map()->m_SelectedEnvelope, SelectedIndex, SelectedChannel, CEditorActionEditEnvelopePointValue::EType::POINT, OldTime, OldValue, CFixedTime::FromSeconds(CurrentTime), f2fx(CurrentValue)));
-
-				if(SelectedIndex != 0)
-				{
-					CurrentTime = pEnvelope->m_vPoints[SelectedIndex].m_Time.AsSeconds();
-				}
-				else
-				{
-					CurrentTime = 0.0f;
-					pEnvelope->m_vPoints[SelectedIndex].m_Time = CFixedTime(0);
-				}
-			}
-
-			s_CurTimeInput.SetFloat(CFixedTime::FromSeconds(CurrentTime).AsSeconds());
-			s_CurValueInput.SetFloat(fx2f(f2fx(CurrentValue)));
-
-			s_CurrentTime = s_CurTimeInput.GetFloat();
-			s_CurrentValue = s_CurValueInput.GetFloat();
-		}
-	}
-
-	View.HSplitTop(6.0f, nullptr, &View);
-	View.HSplitTop(RowHeight, &Row, &View);
-	static int s_DeleteButtonId = 0;
-	const char *pButtonText = pEditor->Map()->IsTangentSelected() ? "Reset" : "Delete";
-	const char *pTooltip = pEditor->Map()->IsTangentSelected() ? "Reset tangent point to default value." : "Delete current envelope point in all channels.";
-	if(pEditor->DoButton_Editor(&s_DeleteButtonId, pButtonText, 0, &Row, BUTTONFLAG_LEFT, pTooltip))
-	{
-		if(pEditor->Map()->IsTangentInSelected())
-		{
-			auto [SelectedIndex, SelectedChannel] = pEditor->Map()->m_SelectedTangentInPoint;
-			const auto &[OldTime, OldValue] = pEditor->Map()->SelectedEnvelopeTimeAndValue();
-			pEditor->Map()->m_EnvelopeEditorHistory.Execute(std::make_shared<CEditorActionResetEnvelopePointTangent>(pEditor->Map(), pEditor->Map()->m_SelectedEnvelope, SelectedIndex, SelectedChannel, true, OldTime, OldValue));
-		}
-		else if(pEditor->Map()->IsTangentOutSelected())
-		{
-			auto [SelectedIndex, SelectedChannel] = pEditor->Map()->m_SelectedTangentOutPoint;
-			const auto &[OldTime, OldValue] = pEditor->Map()->SelectedEnvelopeTimeAndValue();
-			pEditor->Map()->m_EnvelopeEditorHistory.Execute(std::make_shared<CEditorActionResetEnvelopePointTangent>(pEditor->Map(), pEditor->Map()->m_SelectedEnvelope, SelectedIndex, SelectedChannel, false, OldTime, OldValue));
-		}
-		else
-		{
-			auto [SelectedIndex, SelectedChannel] = pEditor->Map()->m_vSelectedEnvelopePoints.front();
-			pEditor->Map()->m_EnvelopeEditorHistory.Execute(std::make_shared<CEditorActionDeleteEnvelopePoint>(pEditor->Map(), pEditor->Map()->m_SelectedEnvelope, SelectedIndex));
-		}
-
-		return CUi::POPUP_CLOSE_CURRENT;
-	}
-
-	return CUi::POPUP_KEEP_OPEN;
-}
-
-CUi::EPopupMenuFunctionResult CEditor::PopupEnvPointMulti(void *pContext, CUIRect View, bool Active)
-{
-	CEditor *pEditor = static_cast<CEditor *>(pContext);
-	const float RowHeight = 12.0f;
-
-	static int s_CurveButtonId = 0;
-	CUIRect CurveButton;
-	View.HSplitTop(RowHeight, &CurveButton, &View);
-	if(pEditor->DoButton_Editor(&s_CurveButtonId, "Project onto", 0, &CurveButton, BUTTONFLAG_LEFT, "Project all selected envelopes onto the curve between the first and last selected envelope."))
-	{
-		static SPopupMenuId s_PopupCurveTypeId;
-		pEditor->Ui()->DoPopupMenu(&s_PopupCurveTypeId, pEditor->Ui()->MouseX(), pEditor->Ui()->MouseY(), 80, 80, pEditor, PopupEnvPointCurveType);
-	}
-
-	return CUi::POPUP_KEEP_OPEN;
-}
-
-CUi::EPopupMenuFunctionResult CEditor::PopupEnvPointCurveType(void *pContext, CUIRect View, bool Active)
-{
-	CEditor *pEditor = static_cast<CEditor *>(pContext);
-	const float RowHeight = 14.0f;
-
-	int CurveType = -1;
-
-	static int s_ButtonLinearId;
-	CUIRect ButtonLinear;
-	View.HSplitTop(RowHeight, &ButtonLinear, &View);
-	if(pEditor->DoButton_MenuItem(&s_ButtonLinearId, "Linear", 0, &ButtonLinear))
-		CurveType = CURVETYPE_LINEAR;
-
-	static int s_ButtonSlowId;
-	CUIRect ButtonSlow;
-	View.HSplitTop(RowHeight, &ButtonSlow, &View);
-	if(pEditor->DoButton_MenuItem(&s_ButtonSlowId, "Slow", 0, &ButtonSlow))
-		CurveType = CURVETYPE_SLOW;
-
-	static int s_ButtonFastId;
-	CUIRect ButtonFast;
-	View.HSplitTop(RowHeight, &ButtonFast, &View);
-	if(pEditor->DoButton_MenuItem(&s_ButtonFastId, "Fast", 0, &ButtonFast))
-		CurveType = CURVETYPE_FAST;
-
-	static int s_ButtonStepId;
-	CUIRect ButtonStep;
-	View.HSplitTop(RowHeight, &ButtonStep, &View);
-	if(pEditor->DoButton_MenuItem(&s_ButtonStepId, "Step", 0, &ButtonStep))
-		CurveType = CURVETYPE_STEP;
-
-	static int s_ButtonSmoothId;
-	CUIRect ButtonSmooth;
-	View.HSplitTop(RowHeight, &ButtonSmooth, &View);
-	if(pEditor->DoButton_MenuItem(&s_ButtonSmoothId, "Smooth", 0, &ButtonSmooth))
-		CurveType = CURVETYPE_SMOOTH;
-
-	std::vector<std::shared_ptr<IEditorAction>> vpActions;
-
-	if(CurveType >= 0)
-	{
-		std::shared_ptr<CEnvelope> pEnvelope = pEditor->Map()->m_vpEnvelopes.at(pEditor->Map()->m_SelectedEnvelope);
-
-		for(int c = 0; c < pEnvelope->GetChannels(); c++)
-		{
-			int FirstSelectedIndex = pEnvelope->m_vPoints.size();
-			int LastSelectedIndex = -1;
-			for(auto [SelectedIndex, SelectedChannel] : pEditor->Map()->m_vSelectedEnvelopePoints)
-			{
-				if(SelectedChannel == c)
-				{
-					FirstSelectedIndex = minimum(FirstSelectedIndex, SelectedIndex);
-					LastSelectedIndex = maximum(LastSelectedIndex, SelectedIndex);
-				}
-			}
-
-			if(FirstSelectedIndex < (int)pEnvelope->m_vPoints.size() && LastSelectedIndex >= 0 && FirstSelectedIndex != LastSelectedIndex)
-			{
-				CEnvPoint FirstPoint = pEnvelope->m_vPoints[FirstSelectedIndex];
-				CEnvPoint LastPoint = pEnvelope->m_vPoints[LastSelectedIndex];
-
-				CEnvelope HelperEnvelope(1);
-				HelperEnvelope.AddPoint(FirstPoint.m_Time, {FirstPoint.m_aValues[c], 0, 0, 0});
-				HelperEnvelope.AddPoint(LastPoint.m_Time, {LastPoint.m_aValues[c], 0, 0, 0});
-				HelperEnvelope.m_vPoints[0].m_Curvetype = CurveType;
-
-				for(auto [SelectedIndex, SelectedChannel] : pEditor->Map()->m_vSelectedEnvelopePoints)
-				{
-					if(SelectedChannel == c)
-					{
-						if(SelectedIndex != FirstSelectedIndex && SelectedIndex != LastSelectedIndex)
-						{
-							CEnvPoint &CurrentPoint = pEnvelope->m_vPoints[SelectedIndex];
-							ColorRGBA Channels = ColorRGBA(0.0f, 0.0f, 0.0f, 0.0f);
-							HelperEnvelope.Eval(CurrentPoint.m_Time.AsSeconds(), Channels, 1);
-							int PrevValue = CurrentPoint.m_aValues[c];
-							CurrentPoint.m_aValues[c] = f2fx(Channels.r);
-							vpActions.push_back(std::make_shared<CEditorActionEnvelopeEditPoint>(pEditor->Map(), pEditor->Map()->m_SelectedEnvelope, SelectedIndex, SelectedChannel, CEditorActionEnvelopeEditPoint::EEditType::VALUE, PrevValue, CurrentPoint.m_aValues[c]));
-						}
-					}
-				}
-			}
-		}
-
-		if(!vpActions.empty())
-		{
-			pEditor->Map()->m_EnvelopeEditorHistory.RecordAction(std::make_shared<CEditorActionBulk>(pEditor->Map(), vpActions, "Project points"));
-		}
-
-		pEditor->Map()->OnModify();
-		return CUi::POPUP_CLOSE_CURRENT;
-	}
-
-	return CUi::POPUP_KEEP_OPEN;
-}
-
 static const auto &&gs_ModifyIndexDeleted = [](int DeletedIndex) {
 	return [DeletedIndex](int *pIndex) {
 		if(*pIndex == DeletedIndex)
@@ -3070,47 +2787,6 @@ CUi::EPopupMenuFunctionResult CEditor::PopupAnimateSettings(void *pContext, CUIR
 	if(pEditor->DoEditBox(&s_SpeedInput, &EditBox, 10.0f, IGraphics::CORNER_NONE, "The animation speed."))
 	{
 		pEditor->m_AnimateSpeed = std::clamp(s_SpeedInput.GetFloat(), MIN_ANIM_SPEED, MAX_ANIM_SPEED);
-	}
-
-	return CUi::POPUP_KEEP_OPEN;
-}
-
-CUi::EPopupMenuFunctionResult CEditor::PopupEnvelopeCurvetype(void *pContext, CUIRect View, bool Active)
-{
-	CEditor *pEditor = static_cast<CEditor *>(pContext);
-
-	if(pEditor->Map()->m_SelectedEnvelope < 0 || pEditor->Map()->m_SelectedEnvelope >= (int)pEditor->Map()->m_vpEnvelopes.size())
-	{
-		return CUi::POPUP_CLOSE_CURRENT;
-	}
-	std::shared_ptr<CEnvelope> pEnvelope = pEditor->Map()->m_vpEnvelopes[pEditor->Map()->m_SelectedEnvelope];
-
-	if(pEditor->m_PopupEnvelopeSelectedPoint < 0 || pEditor->m_PopupEnvelopeSelectedPoint >= (int)pEnvelope->m_vPoints.size())
-	{
-		return CUi::POPUP_CLOSE_CURRENT;
-	}
-	CEnvPoint_runtime &SelectedPoint = pEnvelope->m_vPoints[pEditor->m_PopupEnvelopeSelectedPoint];
-
-	static const char *const TYPE_NAMES[NUM_CURVETYPES] = {"Step", "Linear", "Slow", "Fast", "Smooth", "Bezier"};
-	static char s_aButtonIds[NUM_CURVETYPES] = {0};
-
-	for(int Type = 0; Type < NUM_CURVETYPES; Type++)
-	{
-		CUIRect Button;
-		View.HSplitTop(14.0f, &Button, &View);
-
-		if(pEditor->DoButton_MenuItem(&s_aButtonIds[Type], TYPE_NAMES[Type], Type == SelectedPoint.m_Curvetype, &Button))
-		{
-			const int PrevCurve = SelectedPoint.m_Curvetype;
-			if(PrevCurve != Type)
-			{
-				SelectedPoint.m_Curvetype = Type;
-				pEditor->Map()->m_EnvelopeEditorHistory.RecordAction(std::make_shared<CEditorActionEnvelopeEditPoint>(pEditor->Map(),
-					pEditor->Map()->m_SelectedEnvelope, pEditor->m_PopupEnvelopeSelectedPoint, 0, CEditorActionEnvelopeEditPoint::EEditType::CURVE_TYPE, PrevCurve, SelectedPoint.m_Curvetype));
-				pEditor->Map()->OnModify();
-				return CUi::POPUP_CLOSE_CURRENT;
-			}
-		}
 	}
 
 	return CUi::POPUP_KEEP_OPEN;
